feat:Enhance Marquee Animation with Hover Effects on PopularThisWeek Card #34
