+++ apps/frontend/app/globals.css
@@ -78,3 +78,21 @@
     @apply bg-background text-foreground;
   }
 }
+
+
+@keyframes marquee {
+  0% {
+    transform: translateX(0);
+  }
+  100% {
+    transform: translateX(-50%);
+  }
+}
+
+.animate-marquee {
+  animation: marquee 20s linear infinite;
+}
+
+.animate-pause {
+  animation-play-state: paused;
+}
