+++ apps/frontend/package.json
@@ -24,8 +24,8 @@
     "gsap": "^3.12.7",
     "localforage": "^1.10.0",
     "lucide-react": "^0.446.0",
-    "next": "13.5.1",
-    "postcss": "8.4.30",
+    "next": "^15.2.4",
+    "postcss": "^8.5.3",
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "shadcn-ui": "^0.9.5",
